refactor(frontend): move WalletConnect to lib
