+++ src/frontend/src/eth/components/core/EthListener.svelte
@@ -2,10 +2,10 @@
 	import { isNullish } from '@dfinity/utils';
 	import { onDestroy } from 'svelte';
 	import { initTransactionsListener } from '$eth/services/eth-listener.services';
-	import type { WebSocketListener } from '$eth/types/listener';
 	import { ethAddress } from '$lib/derived/address.derived';
 	import type { OptionEthAddress } from '$lib/types/address';
 	import type { Token } from '$lib/types/token';
+	import type { WebSocketListener } from '$lib/types/listener';
 
 	export let token: Token;
