+++ src/frontend/src/eth/components/wallet-connect/WalletConnectSendModal.svelte
@@ -5,11 +5,11 @@
 	import { enabledEthereumTokens } from '$eth/derived/tokens.derived';
 	import type { EthereumNetwork } from '$eth/types/network';
 	import type {
-		OptionWalletConnectListener,
 		WalletConnectEthSendTransactionParams
 	} from '$eth/types/wallet-connect';
 	import SendTokenContext from '$lib/components/send/SendTokenContext.svelte';
 	import type { Token } from '$lib/types/token';
+	import type { OptionWalletConnectListener } from '$lib/types/wallet-connect';
 
 	export let request: Web3WalletTypes.SessionRequest;
 	export let firstTransaction: WalletConnectEthSendTransactionParams;
