--- src/frontend/src/eth/types/listener.ts
@@ -1,3 +1,0 @@
-export interface WebSocketListener {
-	disconnect: () => Promise<void>;
-}
