+++ src/frontend/src/lib/components/wallet-connect/WalletConnect.svelte
@@ -2,7 +2,7 @@
 	import WalletConnectSend from '$eth/components/wallet-connect/WalletConnectSend.svelte';
 	import WalletConnectSession from '$eth/components/wallet-connect/WalletConnectSession.svelte';
 	import WalletConnectSign from '$eth/components/wallet-connect/WalletConnectSign.svelte';
-	import type { OptionWalletConnectListener } from '$eth/types/wallet-connect';
+	import type { OptionWalletConnectListener } from '$lib/types/wallet-connect';
 
 	let listener: OptionWalletConnectListener;
 </script>
