+++ src/frontend/src/lib/types/listener.ts
@@ -13,3 +13,7 @@ export interface TokenToListener {
 	token: Token;
 	listener: ComponentType;
 }
+
+export interface WebSocketListener {
+	disconnect: () => Promise<void>;
+}
